chore: ensure recipient address exists and is valid + that route is complete
