+++ .changeset/loud-coins-brake.md
@@ -1,0 +1,6 @@
+---
+'@thorswap-lib/swapkit-core': patch
+'@thorswap-lib/types': patch
+---
+
+Typing and address ensure
